+++ packages/abap-deploy-config-inquirer/test/prompts/validators.test.ts
@@ -173,6 +173,7 @@ describe('Test validators', () => {
             expect(PromptState.abapDeployConfig).toStrictEqual({
                 url: 'https://mock.url.target1.com',
                 client: '001',
+                connectPath: undefined,
                 isAbapCloud: false,
                 scp: false,
                 targetSystem: 'https://mock.url.target1.com'
@@ -243,6 +244,7 @@ describe('Test validators', () => {
             expect(result).toBe(true);
             expect(PromptState.abapDeployConfig).toStrictEqual({
                 url: 'https://mock.url.target1.com',
+                connectPath: undefined,
                 client: '001',
                 isAbapCloud: true,
                 scp: true,
@@ -256,6 +258,7 @@ describe('Test validators', () => {
             expect(result).toBe(true);
             expect(PromptState.abapDeployConfig).toStrictEqual({
                 url: 'https://mock.notfound.url.target1.com',
+                connectPath: undefined,
                 isAbapCloud: false,
                 scp: false,
                 targetSystem: undefined,