+++ packages/adp-tooling/src/writer/project-utils.ts
@@ -2,13 +2,7 @@ import { join } from 'node:path';
 import { readFileSync } from 'node:fs';
 import type { Editor } from 'mem-fs-editor';
 
-import {
-    type CloudApp,
-    type AdpWriterConfig,
-    type TypesConfig,
-    type CfAdpWriterConfig,
-    type DescriptorVariant
-} from '../types';
+import type { CloudApp, AdpWriterConfig, TypesConfig, CfAdpWriterConfig, DescriptorVariant } from '../types';
 import {
     enhanceUI5DeployYaml,
     enhanceUI5Yaml,