+++ packages/fiori-mcp-server/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @sap-ux/fiori-mcp-server
 
+## 0.6.45
+
+### Patch Changes
+
+- 55d833f: fix i18next init showSupportNotice: false.
+
+## 0.6.44
+
+### Patch Changes
+
+- 55417bb: fix(deps): update dependency i18next to v25.8.18
+- Updated dependencies [55417bb]
+    - @sap-ux/store@1.5.10
+    - @sap-ux/fiori-docs-embeddings@0.4.16
+
+## 0.6.43
+
+### Patch Changes
+
+- 1b7094e: fix(deps): update dependency @sap/ux-specification to v1.144.0
+
+## 0.6.42
+
+### Patch Changes
+
+- 681f169: fix: MCP server title
+
 ## 0.6.41
 
 ### Patch Changes